Initial checkin for paginated chart options
Previous/next buttons will navigate through the various chart option
blocks, optionally skipping the left axis options if no series has been
selected to be graphed against the left axis.

CSS now needs to be cleaned up, a pagination counter needs to be
implemented, and a "skip to finish" button needs to be implemented to
complete the functionality.

Conflicts:

	index.html
	js/chartbuilder.js